Completion Widget: Move class to his own file
